Changed files structure and updated play-services-maps dependency on Android
