+++ example/App.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import MainNavigator from './navigators/MainNavigator';
 import { Platform } from 'react-native';
-import { Providers } from 'expo-maps/build/Maps.types';
+import { Providers } from 'expo-maps/build/Map.types';
 import SettingsContainer from './components/SettingsContainer';
 import SwitchContainer from './components/SwitchContainer';
 import ProviderContext from './context/ProviderContext';
@@ -27,12 +27,10 @@ export default function App() {
         {Platform.OS === 'ios' && (
           <SettingsContainer style={{ backgroundColor: 'white' }}>
             <SwitchContainer
-              title="Use Apple Maps"
+              title='Use Apple Maps'
               value={provider === 'apple'}
-              onValueChange={() =>
-                setProvider(provider === 'google' ? 'apple' : 'google')
-              }
-              textColor="black"
+              onValueChange={() => setProvider(provider === 'google' ? 'apple' : 'google')}
+              textColor='black'
             />
           </SettingsContainer>
         )}
