+++ expo-maps/android/build.gradle
@@ -75,7 +75,7 @@ repositories {
 
 dependencies {
   implementation project(':expo-modules-core')
-  implementation "com.google.android.gms:play-services-maps:18.0.1"
+  implementation "com.google.android.gms:play-services-maps:18.0.2"
   implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:${safeExtGet('kotlinVersion', '1.4.21')}"
   implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.4.3"
 }
