+++ expo-maps/src/Common.types.ts
@@ -1,0 +1,4 @@
+export type Point = {
+  latitude: number;
+  longitude: number;
+};
